requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
databases[asyncpg]==0.9.0
sqlalchemy==2.0.30
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
pydantic[email]==2.7.1
pydantic-settings==2.2.1
